requirements.txt
pandas==1.5.2
prefect==2.7.4
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.45
